always use double
git-svn-id: https://svn.salilab.org/imp/trunk@6435 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/core/include/internal/evaluate_distance_pair_score.h
@@ -90,7 +90,7 @@ double compute_distance_pair_score(const algebra::VectorD<D> &delta,
                                    algebra::VectorD<D> *d,
                                    SD sd,
                                    double deriv_multiplier = 1.0) {
-  static const Float MIN_DISTANCE = .00001;
+  static const double MIN_DISTANCE = .00001;
   double distance= delta.get_magnitude();
   double shifted_distance = sd(distance);
